password-hash: make `Ident::new` fallible; add `Ident::new_unwrap` (#960)
Unifies the `const fn` implementation of `Ident::new` and `TryFrom`,
with the former now able to return a result thanks to const panic.

Since `Debug` isn't const-friendly yet and therefore `Result::unwrap`
cannot be used in const contexts, adds a `::new_unwrap` function as a
stopgap. It performs some pattern matching and uses const panic to
handle the `Err` case.