elliptic-curve: remove `Zeroize` impl from `ecdh::SharedSecret`
The purpose of zeroization is to clear the bytes of the shared secret on
drop.

However, before this commit `SharedSecret` also had a `Zeroize` impl
which could be called before `Drop`.

Used incorrectly, this could potentially erase the shared secret data,
which could potentially lead to bugs where that data is used to derive
e.g. encryption keys after being zeroed.

This commit removes the `Zeroize` impl to prevent this. Data is still
zeroed on drop and it still impls the `ZeroizeOnDrop` trait.