KEM v2 (#982)
I made a few changes to the KEM API

## Deserialization
As discussed over Zulip earlier this year, there should probably be a way to deserialize an encapsulated key sent over the wire.

This breaks a bit with the rest of RustCrypto because it uses `&GenericArray` rather than `&[u8]` for bitstrings. I'm not married to this, but the pros are:
1. It fits the existing API, which already uses a `GenericArray` to describe shared secrets
2. Fixed size arrays make function preconditions clear a first glance, which is good.

## Shared Secrets
Separately, I've also switched the API off of directly returning `GenericArray`s for a shared secret and made it its own `SharedSecret` type. The con is that the values are just a little more opaque. The pros are:
1. Shared secrets are now zeroize-on-drop
2. The type signatures are more concise