elliptic-curve: generic impl of complete prime order formulas
Adds a generic implementation of the complete addition formulas from
Renes-Costello-Batina 2015[1] adapted from @str4d's original
implementation for the `p256` crate:

https://github.com/RustCrypto/elliptic-curves/pull/15

This implementation has been copied-and-pasted into the `p384` crate,
hence the motivation to make it generic and extract it somewhere that it
can be reused.

The API exposed is fairly low-level, however it's difficult to better
encapsulate it without making breaking changes to the `elliptic-curve`
crate. Thus this PR opts to provide an initial low-level generic
implementation with the goal of exploring removing more duplication with
a higher-level API as followup work to be done at a time when breaking
changes are permitted.

[1]: https://eprint.iacr.org/2015/1060