universal-hash: add back `finalize_reset` (#1054)
This method was used by the `aes-gcm-siv` crate to handle a case where
consuming `self` is problematic and borrowing `&mut self` is much more
straightforward. It seems we've had several cases like this pop up with
similar methods that were removed in other crates like `block-modes` and
`digest`.

The method is provided but bounded on `Clone + Reset`, with the latter
sourced from `crypto-common`.