universal-hash: add `UhfBackend::blocks_needed_to_align`
Originally suggested by @str4d here:

https://github.com/RustCrypto/traits/pull/965#discussion_r857126434

The use case is ensuring unbuffered processing of the input AEAD message
body when the AAD does not line up with the number of blocks it's
processing in parallel.

This number of blocks can be input from the start of the message body to
finish completely filling the buffer, ensuring all subsequent processing
occurs at the start of the buffer.