signature: add `hazmat-preview` feature (#1099)
Adds a `hazmat` module gated under a newly added `hazmat-preview`
feature which calls out the relevant functionality as subject to change
with minor versions.

It adds the following traits:

- `PrehashSigner`
- `PrehashVerifier`

These APIs accept the digest to be signed/verified as a raw byte slice.
This comes with potential misuses like failing to use a
cryptographically secure hash function as the `prehash`, which could
enable existential forgeries of signatures, hence gating it under a
`hazmat-preview` feature and placing it in a `hazmat` module.

Note that we previously explored APIs like this for `DigestSigner`. They
were removed in RustCrypto/signatures#17 due to the afforementioned
misuse potential.

However, these APIs are occasionally needed for implementing protocols
that use special rules for computing hashes (e.g. EIP-712 structured
hashes), or for implementing things like network signing services which
want to accept a prehash of a message to be signed rather than the full
message (to cut down on network bandwidth).

The traits accept a byte slice `prehash`, which permits multiple lengths
and allows the implementation to decide which lengths are valid. This
makes it possible for e.g. ECDSA implementations to automatically
truncate message prehashes which are larger than the field size.