signature_derive: `DigestSigner`/`DigestVerifier` support
Adds support for deriving `DigestSigner`/`DigestVerifier` for types
which impl `hazmat::{PrehashSigner, PrehashVerifier}`.

Custom derive is used instead of a blanket impl in case implementations
wish to specialize the `Digest*` impls. For example, they may want to
use the `Digest` parameter elsewhere, for example as the hash function
HMAC-DRBG is instantiated with for RFC6979.