password-hash: use `Salt` type with `PasswordHasher`
Previously `PasswordHash::hash_password` and `PasswordHash::generate`
took `&str` for `Salt`, however this is a bit of a confusing API because
the first thing it does is attempt to convert to a `Salt` which upholds
several invariants including "B64" encoding.

This change makes the `Salt` parameter explicit so it's clear what type
is responsible for checking those invariants.