elliptic-curve: remove `Reduce` methods
Removes all methods from `Reduce` except the newly renamed
`Reduce::reduce` (formerly `from_uint_reduced`).

This follows #1220 which changed APIs to handle discrepancies between
the serialized size of the big integer type used to represent field
elements and the on-the-wire serialization which respects the size of
the curve's field modulus and order.