elliptic-curve: add (back) reduction from bytes (#1229)
As a replacement for the previous methods on `Reduce` which take bytes
as input, this adds a new curve-specific method for reducing from bytes
to a field element.

Bytes is defined as an associated type rather than inferred from the
unsigned integer parameter. This is because the integer size may not map
exactly to the size of the input, especially for wide reductions.

Bytes are interpreted according to curve-specific rules.