elliptic-curve: fix `Reduce` impls for `NonZeroScalar<C>`
They were still coupled to `FieldBytes` rather than forwarding to
`<Scalar<C> as Reduce<I>>::Bytes` properly.

Also, this commit reverses the relationship so `Reduce` has the non-zero
reduction implementation (forwarding to `Scalar`'s `ReduceNonZero` impl)
and the `ReduceNonZero` impl for `NonZeroScalar<C>` calls `Reduce`.

This is needed because `Reduce` defines the `Bytes` type and
`ReduceNonZero` is bounded on it (in the trait definition).