elliptic-curve: have `serde` feature activate `pkcs8`
Previously the `Serialize`/`Deserialize` impls for `PublicKey` were
gated on both the `pkcs8` and `serde` features being enabled, which is
potentially confusing.

`PublicKey` uses `pkcs8` to serialize the public key as SPKI, which
includes OIDs which identify it as an elliptic curve public key and also
another OID which identifies the curve.

Now that features are namespaced, it's easy to activate `pkcs8` from the
`serde` feature to make this a bit easier.