elliptic-curve: minimum supported `tap` for `bits` is v1.0.1
...when resolving with `-Z minimal-versions`

`tap` is a transitive dependency of `bitvec` by way of `wyz` v0.5, which
misspecifies the required `tap` version. This is addressed in newer v0.6
releases of `wyz`, which `bitvec` does not yet use.

So, as a workaround, this adds an explicit `dep:tap` dependency on
v1.0.1 for the `bits` feature.

This was the last remaining `minimal-versions` incompatibility in this
crate, allowing all features to now be tested in the `minimal-versions`
check in CI.