Migrate to `doc_auto_cfg`
Migrates the `aead`, `cipher`, `crypto-common`, `digest`, and
`universal-hash` crates to use the `doc_auto_cfg` feature which
automatically generates documentation for `cfg`-gated code, which in our
uses is almost exclusively gated on crate features.

This allows the previous manual `doc_cfg` annotations to be removed,
which is nice as they otherwise duplicate work whenever `cfg`-gated code
is used.