(async-)signature: MSRV 1.60
To simplify the workspace, which contains weak/namespaced features, this
bumps both `async-signature` and `signature` to MSRV 1.60.

So as for the bump to have some justification, it adds weak feature
activation for `rand_core?/std`, and uses this impl to write a
`From<rand_core::Error>` conversion which can preserve the original
error as the boxed source.