elliptic-curve: remove references from batch generic params
Previously the traits were generic around a reference type. The problem
with that is reference types have lifetimes, which requires notating
them in generic code, in this case requiring an HRTB per generic
parameter.

HRTBs work because the output has no lifetime dependencies on the input,
but if that's the case we can just explicitly borrow in the trait, which
means no HRTB is needed in generic code, which makes notating the trait
bounds significantly simpler.