elliptic-curve: add `LinearCombinationExt` trait
Adds a trait which is generic around its input parameters for computing
linear combinations.

This allows for different strategies depending on the input parameter
type, for example using stack-allocated intermediates sized to match an
input array, which works on `no_std` targets.

Or, if a slice is passed, intermediate `Vec`s can be used when the
`alloc` feature is available.

Also adds a blanket impl for the old `LinearCombination` trait which
calls the new trait.