Bump `password-hash` to v0.6.0-pre
This is not-for-release, but to signal that we are about to begin making
breaking changes.

Notably it would be nice to finally land #1352, which adds `ShaCrypt`
support and makes the `Encoding` enum `non_exhaustive` so we could
potentially add additional encodings in the future without it being a
breaking change.