elliptic-curve: bump `crypto-bigint` to v0.6.0-pre.8
The latest prerelease of `crypto-bigint` includes changes necessary to
also support heap-backed integers (not that we're planning on using
them), but also includes expanded modular arithmetic support including
an implementation of the Bernstein-Yang modular inversion algorithm.

It also includes a migration to `hybrid-array` which will be necessary
for everything to work together correctly.