digest v0.11.0-pre.9
Due to circular dependencies, it was necessary to temporarily split
`elliptic-curve` out of the workspace. It also can't currently resolve a
solution for Cargo.lock due to `digest`.

The main causes of this are crates like `sha2` and `sha3` which are
circular `dev-dependencies` (i.e. between traits and hashes).

This will temporarily break the build until we can do new releases of
`sha2` and `sha3` as well as crates like `hmac` and `hkdf`. Once these
have been released, we can put everything back into a single workspace.