aead: remove `AeadCore::CiphertextOverhead` (#1737)
This size was intended for AEADs based on padded block cipher modes such
as CBC in order to express the underlying cipher's block size and
therefore the maximum amount of possible padding overhead beyond the
original plaintext, which is a full block (in the case a sentinel block
is added to a block-aligned plaintext input).

However, every AEAD we implement uses counter mode, i.e. a stream cipher
instead of a block cipher, which has no overhead, and as such
`CiphertextOverhead` is set to `U0` in every AEAD implementation we
currently maintain.

Furthermore, to my knowledge there are no standard AEADs which use CBC
or other padded block cipher modes of operation. The original goal was
to support an expired draft specification of a CBC+HMAC AEAD.

Since it doesn't appear to be of use, this PR removes it.