cipher: add traits for tweakable block ciphers (#1721)
The trait design generally follows the `BlockCipherEnc/Dec` traits. We
currently do not have tweakable block cipher implementations which use
the backend capability, but I think it's still worth to include it for
API consistency and to future-proof the traits. Some of the helper
methods (e.g. `encrypt_blocks`) are not translated since it's unclear
how to organize passing of tweaks. Arguably, such methods should not be
used with tweakable block ciphers either way.

As a bridge between tweakable and non-tweakable traits the `ZeroTweak`
wrapper is introduced. It allows users to use tweakable block cipher
implementations with the parts of the ecosystem which expects
non-tweakble block ciphers while still being explicit in the code (e.g.
`ZeroTweak<Threefish256>`).