aead: add explicit nonce API
Building towards a full solution to #1666, this adds an initial API
which supports explicit nonces, implemented as a prefix to the AEAD
message.

Putting the nonce in any other position than the message prefix doesn't
make sense. Nothing else works that way. There are multiple possible
permutations like putting the nonce between the ciphertext, or at the
end, but nobody does that, and the best thing we can do for users is
eliminate unnecessary choices.