Revert "signature: replace `signature_derive` with blanket impls (#1827)" (#1840)
This reverts commit bf4774878c7c2e36451119a0ad24f1db6dc93cd8.

Per #1831 this change breaks inference when there is a single explicit
impl of the `Signer` trait.

It also wasn't possible to add corresponding blanket impls to the
`Async*` traits, e.g. `AsyncSigner` for `AsyncDigestSigner`, because of
the existing blanket impl of `AsyncSigner` for `Signer` which we
definitely want to preserve.

As a general rule of thumb, blanket impls only make sense if they work
100% of the time, which doesn't seem to be happening here.

Closes #1831