signature: remove `derive` feature (#1843)
`signature_derive` was used to provide a `Signer` implementation from a
`DigestSigner` and `PrehashSignature` but it does not appears to be used
in any place I can see.

If this becomes to actually be used, we'll then provide a macro-based
implementation for it but I'll hold on to that until we find a consumer.