Add `type ExpandMsg` to `VoprfParameters` (#1814)
Main motivation was that `decaf448`
(https://github.com/RustCrypto/elliptic-curves/pull/1121) requires
`ExpandMsgXof` instead of `ExpandMsgXmd`.

I was briefly tempted to add a `type Hash` to the `ExpandMsg` trait to
be able to write `type ExpandMsg = ExpandMsg<Hash = Self::Hash>`.
Otherwise it is possible to add a different hash in the `ExpandMsg`
type:
```rust
impl VoprfParamters for Fun128 {
    type Hash = Sha512;
    type ExpandMsg = ExpandMsgXmd<Sha256>;
    ...
}
```
But I thought its a bit much. Let me know if you want me to add that.