elliptic-curve: complete `digest` v0.11.0-rc.0 upgrade
Due to inter-repo dependency relationships and all of the crates
previously using `=` to pin `digest`, some of the features of
`elliptic-curve` had to be disabled to complete the `digest`
v0.11.0-rc.0 release in #1868.

Now that it's been released and there are new compatible versions of the
`hkdf`, `sha2`, and `sha3` crates, it's possible to re-enable this
functionality.