`signature`: add `MultipartSigner` and `MultipartVerifier` (#1880)
This PR adds new traits for multipart messages: `MultipartSigner`,
`RandomizedMultipartSigner`, `RandomizedMultipartSignerMut` and
`MultipartVerifier`.

The idea here is to allow non-contiguous bytes to be passed, which is
necessary when the message has to be constructed from multiple sources
without wanting to allocate memory for a contiguous message. E.g. for
`no_std` environments or when the message is rather big but pre-hashing
is not applicable, e.g. PureEdDSA, ML-DSA or SLH-DSA.

I know this is a rather big breaking change, so let me know what you
think!

These new traits can be implemented by a bunch of crates:
- [x] `ecdsa`: https://github.com/RustCrypto/signatures/pull/982
- [x] `ml-dsa`: https://github.com/RustCrypto/signatures/pull/982
- [x]  `slh-dsa`: https://github.com/RustCrypto/signatures/pull/982
- [x] `bign256`: https://github.com/RustCrypto/elliptic-curves/pull/1221
- [x] `sm2`: https://github.com/RustCrypto/elliptic-curves/pull/1221
- [x] `k256`: https://github.com/RustCrypto/elliptic-curves/pull/1221
- [x] `dsa`: https://github.com/RustCrypto/signatures/pull/982
- [x] `lms`: https://github.com/RustCrypto/signatures/pull/982
- [x] `rsa`: https://github.com/RustCrypto/RSA/pull/525
- [ ] `ed25519-dalek`

Resolves https://github.com/RustCrypto/signatures/issues/959.