elliptic-curve: impl `BatchInvert` for `NonZeroScalar` (#1890)
This implements `BatchInvert` for `NonZeroScalar`.

To accomplish this, I did the following notable things:
- Remove all trait bounds on `BatchInvert` itself.
- Remove `CtOption` from `BatchInvert::batch_invert`s return type.
- Expose `invert_batch_internal()` internally, which now takes an
`invert` function instead of requiring `trait Invert`.
- Implement `MulAssign` for `NonZeroScalar`.

Things that could still be added:
- Mirror all `BatchInvert` implementations on `Scalar`, I left the ones
taking a reference to a slice out.
- Implement `MulAssign` for `NonZeroScalar` on more combinations.

It might be a tiny bit simpler if I implemented `Default` on
`NonZeroScalar`, but that seemed like a footgun to me.