elliptic-curve: re-export `group::Curve` as `CurveGroup` (#1902)
This crate also defines a `Curve` trait, and while that trait is used to
describe an elliptic curve, `group::Curve` is a `Group`, i.e. the
elliptic curve group for a particular curve.

See also: zkcrypto/group#51

To prevent this name clash, this commit re-exports `group::Curve` as
`CurveGroup`.

cc @daxpedda