elliptic-curve: fix SEC1 serialization
It wasn't including the curve OID as `EcParameters`. Previously the
logic for populating this field lived in the `sec1` crate.

Since SEC1 needs a slightly different serialization than PKCS#8 (notably
the latter hoists the curve OID onto `AlgorithmParameters` and omits it
from the SEC1 `EcPrivateKey` this more or less reverts the PKCS#8
serialization logic to what it was before.