Only run CI on latest commit of PR branches. (#1948)
This is a little hack I make use of in my projects and found a bit
annoying when opening PRs here. When successive commits are made to a
PR, cancels CI workflows that are not ran on the latest commit - saving
CI time and making things quicker.

I can also make this PR to other RustCrpyto repos if this is something
we want.