digest: move `EagerHash` from `hmac` (#2014)
Originally added in https://github.com/RustCrypto/MACs/pull/151

The situation is `ecdsa` has a trait for the "default" hash function to
use for a particular curve, which we want to work with RFC6979 and ergo
HMAC, so we would like to use an `EagerHash` bound.

But `rfc6979` is an optional dependency, so we can't depend on its
transitive accessibility to `rfc6979::hmac`. This PR added `hmac` as a
dependency gated on the `digest` feature of `ecdsa`:

https://github.com/RustCrypto/signatures/pull/1076

But as I was looking at `EagerHash` I noticed there's nothing
HMAC-specific about it at all and it can easily be moved to `digest`, so
we're able to use it in bounds without any other dependencies besides
`digest`, which would be nice.