elliptic-curve: add `point::LookupTable` and `BasepointTable` (#2023)
Implementation is originally from the `k256` crate, but written to be
generic over a `Point` type which is bound by the `Group` trait.

Adds a `LookupTable` type which supports precomputed lookup tables for a
given curve point which can be selected from in constant time.

Also adds a feature-gated `BasepointTable` type with a `const fn`
constructor which builds a table for precomputed fixed-base scalar
multiplication using the `Group::generator` as the base point.

The `const fn` support for `BasepointTable` is achieved via a
`LazyLock`-style pattern which computes the table lazily upon first use
in a synchronized fashion. It supports using either
`std::sync::LazyLock`, or for `no_std` targets it also supports using
the `critical-section` feature of the `once_cell` crate.