password-hash: remove `phc::Ident` lifetime (#2107)
`Ident` is a short string with a max length of 32-bytes. Instead of
having a lifetime, we can simply store the 32-bytes as an owned type.

This is an incremental step towards removing the lifetime from the
`phc::PasswordHash` type.

The implementation makes the previous `Buffer` (now `StringBuf`) type
used for `Params` const generic around a particular size, and adds some
basic const init functionality, as well as checks for arithmetic
overflow.