password-hash: remove lifetime from `Salt` type
Following #2107, this also removes the lifetime from `Salt`, but also
changes its internal representation to be raw bytes after "B64"
decoding, with the `SaltString` type handling the "B64" encoded form.

This was the last type with a lifetime on `PasswordHash`, so this also
removes the lifetime from that, as well as the lifetime on methods of
traits including `(Customized)PasswordHasher`.

The `Salt` and `SaltString` types support infallible conversions back
and forth between each other.

Additionally, this changes the `salt` parameter of
`hash_password(_customized)` to be a `&[u8]` rather than a `&str`, with
the password hashing implementation expected to handle "B64" encoding.