password-hash: use `phc` crate
Replaces the `phc` submodule with a re-export of the `phc` crate.

A dependency on the crate is needed to make the blanket impl of
`PasswordVerifier` for `PasswordHasher` work (notably the former is
object safe, and a similar blanket impl can't be supported for MCF since
algorithm-specific rules are required)

This also significantly simplifies the `Error` type now that it's only
responsible for errors involving the traits, rather than errors from
parsing password hashes.