password-hash: relax `FromStr` error bounds (#2120)
Changes the bound on `CustomizedPasswordHasher::Params` to remove the
`Err = Error` bound on `FromStr`.

Instead, the blanket impl of `PasswordVerifier` has been updated to
require `Error: From<E>` where `E` is `FromStr::Err`.

This allows crates to still use custom error types for params, which may
be more expressive for a specific algorithm.