crypto-common: vendor `getrandom::SysRng` (#2135)
It's now been a month since #2096 was merged and there still has not
been a corresponding `getrandom` release.

This commit temporarily internally vendors `SysRng` (which has a very
small implementation) until there is a `getrandom` crate release which
is usable. This will unblock work on rolling out the `Generate` trait.

Once there is, this commit can be reverted.